Move Job interface closer to the client of it
