+++ pkg/jobmanager/manager.go
@@ -20,12 +20,26 @@ import (
 	"github.com/adalton/teleport-exercise/pkg/cgroup/cgroupv1"
 	"github.com/adalton/teleport-exercise/pkg/config"
 	"github.com/adalton/teleport-exercise/pkg/io"
+	"github.com/google/uuid"
 )
 
 const (
 	// Superuser is the name of the user who can access any job.
 	Superuser = "administrator"
 )
+
+// Job defines an interface for objects that behave like jobs.  This enables
+// us to define both a production job type as well a a job type for unit
+// testing.
+type Job interface {
+	Start() error
+	Stop() error
+	Status() *JobStatus
+	StdoutStream() *io.ByteStream
+	StderrStream() *io.ByteStream
+	Name() string
+	Id() uuid.UUID
+}
 
 // JobConstructor is a type that models a function for creating Jobs.
 // This enables us to have a "real" job constructor function as well as
